requirements.txt
# requirements.txt (create this file)
Pillow==10.4.0
django-image-cropping==1.1.0
easy-thumbnails==2.8.2